Moved example VDM check implementation
