+++ tictactoe/adapter-vdm/src/main/kotlin/com/anaplan/engineering/azuki/tictactoe/adapter/vdm/check/GameIsDrawCheck.kt
@@ -3,7 +3,6 @@ package com.anaplan.engineering.azuki.tictactoe.adapter.vdm.check
 import com.anaplan.engineering.azuki.tictactoe.adapter.api.GetPlayOrderBehaviour
 import com.anaplan.engineering.azuki.tictactoe.adapter.vdm.XOModule
 import com.anaplan.engineering.azuki.vdm.DefaultModuleBuilder
-import com.anaplan.engineering.azuki.vdm.DefaultVdmCheck
 
 class GameIsDrawCheck(
     private val gameName: String
