+++ tictactoe/adapter-vdm/src/main/kotlin/com/anaplan/engineering/azuki/tictactoe/adapter/vdm/check/PlayerMoveCountCheck.kt
@@ -4,7 +4,6 @@ import com.anaplan.engineering.azuki.tictactoe.adapter.api.PlayerMoveCountBehavi
 import com.anaplan.engineering.azuki.tictactoe.adapter.vdm.XOModule
 import com.anaplan.engineering.azuki.tictactoe.adapter.vdm.toVdmPlayer
 import com.anaplan.engineering.azuki.vdm.DefaultModuleBuilder
-import com.anaplan.engineering.azuki.vdm.DefaultVdmCheck
 
 class PlayerMoveCountCheck(
     private val gameName: String,
